requirements.txt
django==1.11
paramiko
wmi
pywin32
ptwinrm